refactor: useStarRate 이름 useStarRating으로 변경
